# docs/swagger.yaml
definitions:
  main.album:
    properties:
      artist:
        type: string
      id:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
  models.Interaction:
    properties:
      created_at:
        type: string
      id:
        type: integer
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      settings:
        type: object
      updated_at:
        type: string
    type: object
  models.Message:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      interaction_id:
        type: integer
      role:
        type: string
    type: object
info:
  contact: {}
paths:
  /albums:
    get:
      description: get all albums
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.album'
      summary: Get all albums
    post:
      consumes:
      - application/json
      description: Create new album with input payload
      operationId: create-album
      parameters:
      - description: Album's info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/main.album'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.album'
      summary: Create new album
  /albums/{id}:
    get:
      description: get album by ID
      operationId: get-album-by-id
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.album'
      summary: Get an album by id
  /db/interactions:
    get:
      description: Retrieves a list of all interactions
      operationId: get-all-interactions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Interaction'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all interactions
swagger: "2.0"